# api/openapi.yaml
openapi: 3.1.0
info:
  title: WAM — WhatsApp Marketing API
  version: 0.1.0
  description: |
    WAM — WhatsApp Marketing. Source of truth for HTTP APIs.

    * UI: `GET /api-docs/` (Swagger UI). Spec: `GET /api-docs/openapi.yaml`
      (alias `GET /openapi.yaml`). Legacy `/swagger*` 301s to `/api-docs/`.
    * Dashboard: `/dashboard /connect /contacts /templates /campaigns /analytics`
      (htmx, no React). Labels are embedded in Contacts (`/contacts#labels`), not a separate nav.
    * Connection pairing (QR/phone/logout/status), number check,
      one-off send, contacts/labels CRUD + import, templates CRUD + duplicate,
      campaigns + worker, per-campaign funnel, account overview, CSV export.
    * Persistence: SQLite (`WAM_DB_PATH`), shared with whatsmeow sqlstore.
      Auth: single admin (`WAM_ADMIN_PASSWORD`), session cookie.
    * SaaS path (Postgres `WAM_DATABASE_URL` + `WAM_APP_DATABASE_URL`):
      DB users + RBAC (`admin`/`user`, permission checks per handler),
      per-request org transactions (fail-closed RLS), per-account grants
      (`/api/v1/grants`, enforced in multi-account), signup/verify/reset,
      invites, multi-org sessions with switcher.
  contact:
    name: WAM
  license:
    name: UNLICENSED
    identifier: UNLICENSED
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://wam.example.com
    description: Production
...
paths:
...
  /api/v1/contacts:
    get:
      tags: [Contacts]
      operationId: listContacts
      summary: List contacts
      parameters:
        - $ref: '#/components/parameters/PageLimit'
        - $ref: '#/components/parameters/PageCursor'
        - name: group_id
          in: query
          schema: { type: string, format: uuid }
        - name: q
          in: query
          schema: { type: string }
      responses:
        '200':
          description: A page of contacts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContactPage'
    post:
      tags: [Contacts]
      operationId: createContact
      summary: Create contact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContactCreate'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contact'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Phone already exists
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
...
  /api/v1/campaigns/{id}:
    get:
      tags: [Campaigns]
      operationId: getCampaign
      summary: Get campaign by id
      parameters:
        - $ref: '#/components/parameters/EntityId'
      responses:
        '200':
          description: The campaign
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Campaign'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Campaigns]
      operationId: updateCampaign
      summary: Update draft/scheduled campaign
      parameters: [{ $ref: '#/components/parameters/EntityId' }]
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CampaignCreate'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Campaign'
        '400':
          description: Not editable in current status
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Campaigns]
      operationId: deleteCampaign
      summary: Delete campaign with recipients
      parameters: [{ $ref: '#/components/parameters/EntityId' }]
      responses:
        '204': { description: Deleted }
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  parameters:
    EntityId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    PageLimit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    PageCursor:
      name: cursor
      in: query
      schema: { type: string }
  schemas:
...
    Problem:
      title: RFC 9457 problem+json
      type: object
      required: [title, status]
      properties:
        type: { type: string, format: uri, default: 'about:blank' }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
...
    Campaign:
      type: object
      required: [id, name, status, createdAt]
      properties:
        id: { type: string, format: uuid }
        name: { type: string, minLength: 1, maxLength: 200 }
        status: { type: string, enum: [draft, scheduled, sending, paused, done, cancelled] }
        audience:
          type: object
          properties:
            group_ids:
              type: array
              items: { type: string, format: uuid }
            contact_ids:
              type: array
              items: { type: string, format: uuid }
        bodyTemplate: { type: string }
        scheduledAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
    CampaignCreate:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 200 }
        bodyTemplate: { type: string, description: 'Supports {{name}} variable' }
        template_id: { type: string, format: uuid, description: 'Optional template to copy body from when bodyTemplate is empty' }
        group_ids:
          type: array
          items: { type: string, format: uuid }
        contact_ids:
          type: array
          items: { type: string, format: uuid }
        scheduledAt: { type: string, format: date-time, nullable: true }
...
    Contact:
      type: object
      required: [id, phone, createdAt]
      properties:
        id: { type: string, format: uuid }
        phone: { type: string, description: E.164 phone number, example: '+15551234567' }
        name: { type: string, nullable: true }
        waOk: { type: boolean, nullable: true }
        groups:
          type: array
          items: { $ref: '#/components/schemas/Group' }
        createdAt: { type: string, format: date-time }
    ContactCreate:
      type: object
      required: [phone]
      properties:
        phone: { type: string, example: '+15551234567' }
        name: { type: string }
        group_ids:
          type: array
          items: { type: string, format: uuid }
    ContactPage:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/Contact' }
        nextCursor: { type: string, nullable: true }
...
    Group:
      type: object
      required: [id, name, contactCount]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        color: { type: string, nullable: true }
        contactCount: { type: integer }
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: Not found
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...